+++ src/App.tsx
@@ -4,11 +4,11 @@ import { ChakraProvider, useToast } from '@chakra-ui/react'
 import { QueryClientProvider } from '@tanstack/react-query'
 
 import { queryClient } from '@/api/instance'
-import { Routes } from '@/routes'
 import theme from '@/styles/theme'
 
 import { CookieAlaram } from './components/CookieAlaram'
 import { useSSEMessage } from './hooks/useSSEMessage'
+import { Routes } from './routes'
 
 function App() {
   const { message } = useSSEMessage()
